requirements.txt
imageio==2.8.0
matplotlib==3.1.3
numpy==1.18.1
scipy==1.4.1
Keras==2.3.1
scikit_image==0.16.2
opencv_contrib_python==4.1.2.30
Pillow==7.1.2
scikit_learn==0.23.1
skimage==0.0
